Expand function capacity
The function reduceAndMerge did not allow 'family_members' to be displayed correctly in the excel file (clinical-data report). All values were merged into one row instead of being split into multiple rows (on row per value). The function was expecting a certain structure of the data in order to explore it and manipulate it. I tweaked it a little bit to take into account a more complex structure.